third_party/WebKit/LayoutTests/http/tests/devtools/elements/styles-1/color-swatch-expected.txt

The patch verifies that color swatch functions properly in matched and computed styles. crbug.com/461363


Running: selectNode

Running: testColorSwatchInMatchRules
Initial color value: red
After shift-click: rgb(255, 0, 0)
Has popover before click: false
Has popover after click: true

Running: testColorSwatchInCustomProperty
Custom property has a color swatch: true

Running: testColorSwatchInVarFunction
var function has a color swatch: true

Running: testColorSwatchInComputedRules
Initial color value: rgb(255, 0, 0)
After shift-click color value: hsl(0, 100%, 50%)
